Finished Registration page and identity service for now

+++ car-portal-sales/src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -42,11 +43,14 @@ public class AuctionsController : ControllerBase
         if(auction == null) return NotFound();
         return _mapper.Map<AuctionDTO>(auction);
     }
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDTO>> CreateAuction(CreateAuctionDTO auctionDTO) 
     {
         var auction = _mapper.Map<Auction>(auctionDTO);
+
         // TODO : add current user as seller
+        auction.Seller = User.Identity!.Name;
 
         _context.Auctions.Add(auction);
 
@@ -63,12 +67,16 @@ public class AuctionsController : ControllerBase
 
         return CreatedAtAction(nameof(GetAuctionById), new {auction.Id}, newAuction);
     }
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDTO updateAuctionDTO) 
     {
         var auction = await _context.Auctions.Include(x => x.Item).FirstOrDefaultAsync(x => x.Id == id);
         if(auction == null) return NotFound();
+
         // TODO : check seller == username
+        if(auction.Seller != User.Identity!.Name) return Forbid();
+
         auction.Item!.Make = updateAuctionDTO.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDTO.Model ?? auction.Item.Model;
         auction.Item.Color = updateAuctionDTO.Color ?? auction.Item.Color;
@@ -81,12 +89,16 @@ public class AuctionsController : ControllerBase
         if(!result) return BadRequest("Could not save changes to DB");
         return Ok();
     }
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id) 
     {
         var auction = await _context.Auctions.FindAsync(id);
         if(auction == null) return NotFound();
+
         // TODO : check seller == username
+        if(auction.Seller != User.Identity!.Name) return Forbid();
+
         _context.Auctions.Remove(auction);
 
         await _publishEndpoint.Publish<AuctionDeleted>(new
